start change global regression to section regression

+++ face3000/liblinear/linear.h
@@ -9,15 +9,15 @@ extern "C" {
 struct feature_node
 {
 	int index;
-	double value;
+	float value;
 };
 
 struct problem
 {
 	int l, n;
-	double *y;
+	float *y;
 	struct feature_node **x;
-	double bias;            /* < 0 if no bias term */  
+	float bias;            /* < 0 if no bias term */  
 };
 
 enum { L2R_LR, L2R_L2LOSS_SVC_DUAL, L2R_L2LOSS_SVC, L2R_L1LOSS_SVC_DUAL, MCSVM_CS, L1R_L2LOSS_SVC, L1R_LR, L2R_LR_DUAL, L2R_L2LOSS_SVR = 11, L2R_L2LOSS_SVR_DUAL, L2R_L1LOSS_SVR_DUAL }; /* solver_type */
@@ -27,12 +27,12 @@ struct parameter
 	int solver_type;
 
 	/* these are for training only */
-	double eps;	        /* stopping criteria */
-	double C;
+	float eps;	        /* stopping criteria */
+	float C;
 	int nr_weight;
 	int *weight_label;
-	double* weight;
-	double p;
+	float* weight;
+	float p;
 };
 
 struct model
@@ -40,17 +40,17 @@ struct model
 	struct parameter param;
 	int nr_class;		/* number of classes */
 	int nr_feature;
-	double *w;
+	float *w;
 	int *label;		/* label of each class */
-	double bias;
+	float bias;
 };
 
 struct model* train(const struct problem *prob, const struct parameter *param);
-void cross_validation(const struct problem *prob, const struct parameter *param, int nr_fold, double *target);
+void cross_validation(const struct problem *prob, const struct parameter *param, int nr_fold, float *target);
 
-double predict_values(const struct model *model_, const struct feature_node *x, double* dec_values);
-double predict(const struct model *model_, const struct feature_node *x);
-double predict_probability(const struct model *model_, const struct feature_node *x, double* prob_estimates);
+float predict_values(const struct model *model_, const struct feature_node *x, float* dec_values);
+float predict(const struct model *model_, const struct feature_node *x);
+float predict_probability(const struct model *model_, const struct feature_node *x, float* prob_estimates);
 
 int save_model(const char *model_file_name, const struct model *model_);
 struct model *load_model(const char *model_file_name);
@@ -62,8 +62,8 @@ struct model *load_model_bin(std::ifstream& fin);
 int get_nr_feature(const struct model *model_);
 int get_nr_class(const struct model *model_);
 void get_labels(const struct model *model_, int* label);
-double get_decfun_coef(const struct model *model_, int feat_idx, int label_idx);
-double get_decfun_bias(const struct model *model_, int label_idx);
+float get_decfun_coef(const struct model *model_, int feat_idx, int label_idx);
+float get_decfun_bias(const struct model *model_, int label_idx);
 
 void free_model_content(struct model *model_ptr);
 void free_and_destroy_model(struct model **model_ptr_ptr);
