+++ face3000/liblinear/tron.h
@@ -4,9 +4,9 @@
 class function
 {
 public:
-	virtual double fun(double *w) = 0 ;
-	virtual void grad(double *w, double *g) = 0 ;
-	virtual void Hv(double *s, double *Hs) = 0 ;
+	virtual float fun(float *w) = 0 ;
+	virtual void grad(float *w, float *g) = 0 ;
+	virtual void Hv(float *s, float *Hs) = 0 ;
 
 	virtual int get_nr_variable(void) = 0 ;
 	virtual ~function(void){}
@@ -15,17 +15,17 @@ public:
 class TRON
 {
 public:
-	TRON(const function *fun_obj, double eps = 0.1, int max_iter = 1000);
+	TRON(const function *fun_obj, float eps = 0.1, int max_iter = 1000);
 	~TRON();
 
-	void tron(double *w);
+	void tron(float *w);
 	void set_print_string(void (*i_print) (const char *buf));
 
 private:
-	int trcg(double delta, double *g, double *s, double *r);
-	double norm_inf(int n, double *x);
+	int trcg(float delta, float *g, float *s, float *r);
+	float norm_inf(int n, float *x);
 
-	double eps;
+	float eps;
 	int max_iter;
 	function *fun_obj;
 	void info(const char *fmt,...);
