+++ face3000/randomforest.h
@@ -10,11 +10,11 @@ public:
 	int samples_;
 	bool is_leaf_;
 	int depth_; // recording current depth
-	double threshold_;
+	float threshold_;
 	bool thre_changed_;
 	FeatureLocations feature_locations_;
-	Node(Node* left, Node* right, double thres, bool leaf);
-	Node(Node* left, Node* right, double thres);
+	Node(Node* left, Node* right, float thres, bool leaf);
+	Node(Node* left, Node* right, float thres);
 	Node();
 };
 
@@ -25,34 +25,34 @@ public:
 	int landmark_index_;
 	int tree_depth_;
 	int trees_num_per_forest_;
-	double local_radius_;
+	float local_radius_;
 	int all_leaf_nodes_;
-	//cv::Mat_<double> mean_shape_;
+	//cv::Mat_<float> mean_shape_;
 	std::vector<Node*> trees_;
 	std::vector<FeatureLocations> local_position_; // size = param_.local_features_num
-	std::vector<cv::Mat_<double> >* regression_targets_;
+	std::vector<cv::Mat_<float> >* regression_targets_;
 
-	bool TrainForest(//std::vector<cv::Mat_<double> >& regression_targets, 
+	bool TrainForest(//std::vector<cv::Mat_<float> >& regression_targets, 
 		const std::vector<cv::Mat_<uchar> >& images,
 		const std::vector<int>& augmented_images_index,
-		//const std::vector<cv::Mat_<double> >& augmented_ground_truth_shapes,
+		//const std::vector<cv::Mat_<float> >& augmented_ground_truth_shapes,
 		const std::vector<BoundingBox>& augmented_bboxes,
-		const std::vector<cv::Mat_<double> >& augmented_current_shapes,
-		const std::vector<cv::Mat_<double> >& rotations,
-		const std::vector<double>& scales);
+		const std::vector<cv::Mat_<float> >& augmented_current_shapes,
+		const std::vector<cv::Mat_<float> >& rotations,
+		const std::vector<float>& scales);
 	Node* BuildTree(std::set<int>& selected_indexes, cv::Mat_<int>& pixel_differences, std::vector<int>& images_indexes, int current_depth);
 	int FindSplitFeature(Node* node, std::set<int>& selected_indexes,
 		cv::Mat_<int>& pixel_differences, std::vector<int>& images_indexes, std::vector<int>& left_indexes, std::vector<int>& right_indexes);
-	cv::Mat_<double> GetBinaryFeatures(const cv::Mat_<double>& image,
-		const BoundingBox& bbox, const cv::Mat_<double>& current_shape, const cv::Mat_<double>& rotation, const double& scale);
+	cv::Mat_<float> GetBinaryFeatures(const cv::Mat_<float>& image,
+		const BoundingBox& bbox, const cv::Mat_<float>& current_shape, const cv::Mat_<float>& rotation, const float& scale);
 	int MarkLeafIdentity(Node* node, int count);
-	int GetNodeOutput(Node* node, const cv::Mat_<double>& image,
-		const BoundingBox& bbox, const cv::Mat_<double>& current_shape, const cv::Mat_<double>& rotation, const double& scale);
+	int GetNodeOutput(Node* node, const cv::Mat_<float>& image,
+		const BoundingBox& bbox, const cv::Mat_<float>& current_shape, const cv::Mat_<float>& rotation, const float& scale);
 	//predict()
-	int GetBinaryFeatureIndex(int tree_index, const cv::Mat_<double>& image,
-	const BoundingBox& bbox, const cv::Mat_<double>& current_shape, const cv::Mat_<double>& rotation, const double& scale);
+	int GetBinaryFeatureIndex(int tree_index, const cv::Mat_<float>& image,
+	const BoundingBox& bbox, const cv::Mat_<float>& current_shape, const cv::Mat_<float>& rotation, const float& scale);
 	RandomForest();
-	RandomForest(Parameters& param, int landmark_index, int stage, std::vector<cv::Mat_<double> >& regression_targets);
+	RandomForest(Parameters& param, int landmark_index, int stage, std::vector<cv::Mat_<float> >& regression_targets);
 	void WriteTree(Node* p, std::ofstream& fout);
 	Node* ReadTree(std::ifstream& fin);
 	void SaveRandomForest(std::ofstream& fout);
